round correctly when there was precision that was truncated.
When a number such as 1.01 was asked to be rounded to precision 0, we
were not rounding correctly because we only looked 1 decimal place back
and saw a zero there. Fix this by not calling the rounding methods at
all if there is no other precision available and using the knowledge to
know when to round up from 1.01.

Added a bunch of test numbers to test this and other potential issue
cases.

Fixes #1